+++ src/const.js
@@ -4,5 +4,10 @@ const EVENT_TYPE = [
 const EVENT_DESTINATION = [
   `Chamonix`, `Geneva`, `Amsterdam`, `Helsinki`, `Oslo`, `Kopenhagen`, `Den Haag`, `Rotterdam`, `Saint Petersburg`, `Moscow`, `Sochi`, `Tokio`, `Kioto`, `Nagasaki`, `Hiroshima`, `Berlin`, `Munich`, `Frankfurt`, `Vien`, `Rome`, `Naples`, `Venice`, `Milan`, `Monaco`, `Paris`, `Barcelona`, `Valencia`, `Madrid`
 ];
+const RANDOM_PRICE_SETTINGS = {
+  MIN_PRICE: 1000,
+  MAX_PRICE: 10000,
+  MULTIPLE: 10,
+};
 
-export {EVENT_TYPE, EVENT_DESTINATION};
+export {EVENT_TYPE, EVENT_DESTINATION, RANDOM_PRICE_SETTINGS};
